MeanFieldTheory

Documentation for MeanFieldTheory.

MeanFieldTheory.PureTBA

MeanFieldTheory.SCMF

MeanFieldTheory.constant

MeanFieldTheory.meanfield

MeanFieldTheory.normal

MeanFieldTheory.Ω

MeanFieldTheory.Ω₀

TightBindingApproximation.Fitting.optimize!

MeanFieldTheory.PureTBA — Type

const PureTBA{K<:TBAKind} = Union{SimpleTBA{K}, CompositeTBA{K}}

Pure tight-binding-approximation, type alias for Union{SimpleTBA{K}, CompositeTBA{K}}.

MeanFieldTheory.SCMF — Type

SCMF{K<:Fermionic, N<:PureTBA{K}, B<:BrillouinZone, I<:Parameters, M<:PureTBA{K}, C<:Function} <: TBA{K, N, Nothing}

Self-consistent mean-field theory for fermionic systems.

MeanFieldTheory.constant — Method

constant(scmf::SCMF) -> Real
constant(scmf::Algorithm{<:SCMF}) -> Real

Get the constant part of the free energy of a fermionic system at the mean-field level.

MeanFieldTheory.meanfield — Method

meanfield(scmf::SCMF) -> TBA
meanfield(scmf::Algorithm{<:SCMF}) -> TBA

Get the mean-field part.

MeanFieldTheory.normal — Method

normal(scmf::SCMF) -> TBA
normal(scmf::Algorithm{<:SCMF}) -> TBA

Get the normal part.

MeanFieldTheory.Ω — Method

Ω(scmf::SCMF; kwargs...) -> Real
Ω(scmf::Algorithm{<:SCMF}; kwargs...) -> Real

Get the free energy of a fermionic system at the mean-field level.

MeanFieldTheory.Ω₀ — Method

Ω₀(scmf::SCMF; kwargs...) -> Real
Ω₀(scmf::Algorithm{<:SCMF}; kwargs...) -> Real

Get the free energy of a fermionic system of the normal state at the mean-field level.

TightBindingApproximation.Fitting.optimize! — Function

optimize!(
    scmf::Union{SCMF, Algorithm{<:SCMF}}, variables=keys(Parameters(meanfield(scmf)));
    verbose=false,
    method=LBFGS(),
    options=Options(x_abstol=4*10^-6, f_abstol=4*10^-6, iterations=1000, show_trace=true),
    condensation::Bool=false,
    kwargs...
)

Optimize the order parameters of a fermionic system by the self-consistent mean-field theory.